Split test_monitor into two modules: test_monitor_duration_elapsed and test_monitor_interrupted.
